Allow URI to be set on init
My DeepL key requires V2 of the API, though the existing calls seem to work with it.
Note that V2 has a separate URI for free vs paid plans.
Added a param allow for provision of a URI on init. We're unable to do this after init due to the call to `getLanguages()`.